Use two endpoints to get Task & Container metadata
`Limits` is an optional field in both the container response, i.e. what
is returned from the ECS_CONTAINER_METADATA_URI_V4, and also in the task
response, i.e. ECS_CONTAINER_METADTA_URI_V4/task. Instead of failing
when the Task limits are missing, both endpoints are used to reconcile
the appropriate limit for GOMAXPROCS

- removes Container ID from the Config struct. the last piece of the URI
  seemed to have no relation to the container or tasks. uses the
  DockerId field in both responses to find matching container
- interrogates ECS_CONTAINER_METADATA_URI_V4 and unmarshals to the
  Container struct. uses new Task method getContainerMeta() for this.
- sanity check that both responses do not return 0 (empty) Limits
- sanity check that the container limit is not greater than the task
  limit
- adds dedicated task_test tests for handling endpoint errors and any
  errors raised when trying to process Limits